Update 05-04-2020 at 10.32 pm

+++ backend/modules/content/controllers/ContactController.php
@@ -188,7 +188,11 @@ class ContactController extends Controller
                     $model->main_photo = $path;
                     $file->saveAs('uploads/contact/' . $path);
                     if (isset($old_name)) {
-                        unlink('uploads/contact/' . $old_name);
+                        if($old_name==''){
+
+                        }else {
+                            unlink('uploads/contact/' . $old_name);
+                        }
                     } else {
                         // Do nothing
                     }
@@ -210,6 +214,7 @@ class ContactController extends Controller
                             $modelDetail->photo_url = $path;
                             ${'profile_file' . $c}->saveAs('uploads/contact/related_photo/' . $path);
                             if (isset($old_name)) {
+
                                 unlink('uploads/contact/related_photo/' . $old_name);
                             } else {
                                 // Do nothing
